updated product list to cards
