+++ src/app.scss
@@ -15,3 +15,7 @@ a {
 	color: inherit;
 	text-decoration: none;
 }
+
+ul {
+	list-style: none;
+}
