+++ src/pages/home/Home.tsx
@@ -1,16 +1,25 @@
-import ProductShowCard from "../../components/ProductShowCard"
+import ProductShowCard from "./ProductShowCard"
 import { productList } from "./productSpecs"
+import { Box } from "@mui/material"
 const Home = () => {
 	return (
 		<div>
 			<h1>Welcome to Free INV!</h1>
 			<p>Your FREE inventory management application</p>
 			<p>{`Tap any subscription below to sign up -->`}</p>
-			<div>
+			<Box
+				sx={{
+					display: "flex",
+					flexWrap: "wrap",
+					alignItems: "center",
+					justifyContent: "center",
+					gap: 2,
+				}}
+			>
 				{productList.map((product) => (
 					<ProductShowCard key={product.title} productSpec={product} />
 				))}
-			</div>
+			</Box>
 		</div>
 	)
 }
